5683 populate collapsible on filter summary
